refactor: Moved logging to separate module and cleaned out println macros
